PRACTICAL 4
First, run the document 'extract_variable.py', it is a function to extract any variables in a specific dataset. Then move to 'Practical4.py 'and run it, there's 8 datasets in practical4, with the function we just ran above, we can now easily extract the variables from any of these provided datasets.