update 11.3
For EVS_code_1 I don't know why scipy is not working when I'm importing this to VS code. This code explore the precipitation data from Beijing, Shijiazhuang and Chengde, and generate some figures to see the trends of precipitation change.
For 11.3 PRACTICAL I use Chatgpt to remove the the blank areas on the edge part of figure and explain the error.